requirements.txt
numpy
opencv-python
Pillow
ultralytics
torch
scikit-learn
tk
matplotlib